backend/requirement.txt
Flask
PyJWT
Werkzeug
Flask-CORS
Flask-SQLAlchemy
joblib
lightgbm
numpy
scikit-learn